requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
google-generativeai==0.8.3
pillow==11.0.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiofiles==24.1.0
